Reorganize everything
