+++ bin/print-manifest.js
