+++ package.json
@@ -4,20 +4,20 @@
   "description": "React Native survey app, together with supporting infrastructure.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "vitest run"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
-  "workspaces": {
-    "packages": [
-      "packages/fika-collect-lambda",
-      "packages/fika-collect-survey-editor",
-      "packages/fika-collect-survey-schema"
-    ],
-    "nohoist": [
-      "**/react-refresh",
-      "**/react-refresh/**"
-    ]
+  "dependencies": {
+    "locale-codes": "^1.3.1",
+    "typescript": "^4.9.5",
+    "zod": "^3.24.3"
+  },
+  "devDependencies": {
+    "@types/tape": "^5.8.1",
+    "glob": "^11.0.1",
+    "ts-node": "^10.9.2",
+    "vitest": "^3.1.4"
   }
 }
