+++ src/editor/README.md
