+++ src/editor/public/manifest.json
