+++ src/editor/vite.config.js
