+++ test/schema/fixtures/detailed_report.json
