+++ vitest.config.mjs
@@ -2,7 +2,7 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
   test: {
-    include: ["test/*.test.ts"],
+    include: ["test/**/*.test.ts"],
     globals: true,
     environment: "node",
   },
